+++ test/fuzz/asset-modules/StargateAssetModule/GetUnderlyingAssets.fuzz.t.sol
@@ -22,7 +22,7 @@ contract GetUnderlyingAssets_StargateAssetModule_Fuzz_Test is StargateAssetModul
                               TESTS
     /////////////////////////////////////////////////////////////// */
 
-    function testFuzz_Success_getUnderlyingAssets(uint96 positionId, uint256 poolId) public {
+    function testFuzz_Success_getUnderlyingAssets(uint96 positionId, uint96 poolId) public {
         // Given : The underlying token of the pool is an asset added to the Registry
         poolMock.setToken(address(mockERC20.token1));
